brain/requirements.txt
fastapi
uvicorn
httpx
paho-mqtt
python-dotenv
anthropic>=0.18.0
groq>=0.4.0
prometheus-client
python-kasa
pywizlight
tzdata
